tsc error fixes
