+++ packages/desktop-client/src/components/AnimatedRefresh.tsx
@@ -23,7 +23,9 @@ export default function AnimatedRefresh({
 }: AnimatedRefreshProps) {
   return (
     <View
-      style={{ animation: animating ? `${spin} 1s infinite linear` : null }}
+      style={{
+        animation: animating ? `${spin} 1s infinite linear` : undefined,
+      }}
     >
       <Refresh width={14} height={14} style={iconStyle} />
     </View>
