+++ packages/desktop-client/src/components/sort.tsx
@@ -3,38 +3,16 @@ import React, {
   useEffect,
   useRef,
   useLayoutEffect,
-  useMemo,
   useState,
   useContext,
-  type RefCallback,
-  type MutableRefObject,
   type Context,
-  type Ref,
 } from 'react';
 import { useDrag, useDrop } from 'react-dnd';
 
+import { useMergedRefs } from '../hooks/useMergedRefs';
 import { theme } from '../style';
 
 import View from './common/View';
-
-function useMergedRefs<T>(
-  ref1: RefCallback<T> | MutableRefObject<T>,
-  ref2: RefCallback<T> | MutableRefObject<T>,
-): Ref<T> {
-  return useMemo(() => {
-    function ref(value) {
-      [ref1, ref2].forEach(ref => {
-        if (typeof ref === 'function') {
-          ref(value);
-        } else if (ref != null) {
-          ref.current = value;
-        }
-      });
-    }
-
-    return ref;
-  }, [ref1, ref2]);
-}
 
 type DragState = {
   state: 'start-preview' | 'start' | 'end';
