+++ packages/desktop-client/src/hooks/useMergedRefs.ts
@@ -1,0 +1,25 @@
+import {
+  type MutableRefObject,
+  type Ref,
+  type RefCallback,
+  useMemo,
+} from 'react';
+
+export function useMergedRefs<T>(
+  ref1: RefCallback<T> | MutableRefObject<T>,
+  ref2: RefCallback<T> | MutableRefObject<T>,
+): Ref<T> {
+  return useMemo(() => {
+    function ref(value: T) {
+      [ref1, ref2].forEach(ref => {
+        if (typeof ref === 'function') {
+          ref(value);
+        } else if (ref != null) {
+          ref.current = value;
+        }
+      });
+    }
+
+    return ref;
+  }, [ref1, ref2]);
+}
